test: move testutil to testdefault
